user story #2970024 - refactored code
